extracted createCapabilityFiltersContext to a seperate file
